Add MIT license, mention used-styles library and move files to subfolder
