+++ packages/react/src/ssr/used-styles-hds/loadStyleDefinitions.ts
@@ -1,3 +1,7 @@
+/**
+ * The code here is based on used-styles npm library: https://github.com/theKashey/used-styles
+ * Some of the code was leaking to browser causing problems, so we copied the code instead of using the library as is.
+ */
 import { AtRule as PostcssAtRule, parse, Rule } from 'postcss';
 import { str as crc32Str } from 'crc-32';
